add interactive visualization to the selection-based approach to problem 0001
wrestled a little with MathJAX vs v-model input elements... but it was a distraction, the real problem was I was using both :value and a default value for the ref(...) that backed the v-model.

oh, also reorganized the content a little bit.  going forward, the <script setup> block will always be at the top and the main <style> block will be at the very end (post-content).  This makes it easy to locate either one within the source view of the page.  I will move to behavior-oriented external .js files if the scripting content gets too high (but this one has a short scripting block even with a generalized solution duplicated in text and code)

500 lines seems to be a reasonable target for approximately done, however future episodes will not likely have so much variation.