include footnote processing for the blog posts
it appears that these notes are always footnotes and not endnotes (at the end of the current section) and this is not easily fixed with markdown-it-footnote config or inline annotations.  The result is a little confusing with the exercises also appearing as numbered lists right above, and I don't care to have a "footnotes" section show up in the table of contents.

it's not really worth fussing over much, especially since the only problem using it is still in a draft stage right now