reorganize golang solutions to be in the same package.
I had suspected there would be some benefit to having each in its own package (they typically don't depend on each other and only sometimes need to access a common utility in a sibling package)..

One disadvantage to this is that when scanning for provided solutions (in the data-loading code of Vue/Vitepress) it has to visit each directory (if the name matches p####) and then check that a solve.go exists there... but this differs from the other solution collections in other languages, and for no good reason.  So let's reorg before this gets too numerous.