fix a couple of display bugs and tweak the Prolog code example
seems there's interfereince in getting a snippet region in Prolog, despite trying many comment styles and region naming.. works fine for golang, so I have the syntax right when the comments are // but even with `% // #region...` it isn't working.

Not worth digging into, tbh, I'll just duplicate the intended region because it's not likely to change in the future.