ocaml solution for problem 0001
I've been wanting to include a functional language in this set for a little while.. after going back and forth between Lisp and Haskell I decided I would actually like to hone my Ocaml skills and it is a fairly good representation of functional languages even if it doesn't benefit from certain optimizations that GHC does easily (e.g. deforesting) or the same expressiveness as Lisp (its macro system and CLOS are slightly better than ocaml's, imo, but then I'm still learning ocaml).  I think ocaml is easier to read than both Lisp and Haskell, especially for the uninitiated, and this site is all about professing for the learning programmer.