update README
I think we already have good coverage of language philosophies, now that ocaml is being included and with the possible inclusion of Lua for a scripting language (but Python kind of covers that already, too, but Lua is a little less batteries-included).  Cue may need to be cut, it's kind of a stretch but it makes for an interesting challenge.  I suppose a language with heavier object-oriented features (like Java or C#) could be added but I'm not going to split my focus on that.